Add option to not remove feature prefixes (for EG genomes)
EG sometimes have genes and transcripts with the same id, so they add
the biotype as a prefix in the GFF3 id because the same id can't be used
for different biotypes in the GFF3 specification.

By default the pipeline removes those prefix (for exporting data from
genomes not from EG/Ensembl), by to also test the dumper for those
genomes I've added this option so that the final GFF3 passes validation.